refactor(calm-hub): move requireNamespace onto NamespaceStore as a default method
Replace the standalone NamespaceGuard utility with a default method,
NamespaceStore#requireNamespace, so call sites read as
namespaceStore.requireNamespace(x) instead of threading the
NamespaceStore instance through a separate static helper.

Store-level test mocks were intercepting the default method itself
(Mockito stubs interface default methods to a no-op unless told
otherwise), so each store test's setup now opts the mocked
NamespaceStore into running the real requireNamespace body via
doCallRealMethod(), which in turn defers to the already-mocked
namespaceExists().
