+++ calm-hub/src/main/java/org/finos/calm/store/mongo/MongoArchitectureStore.java
@@ -13,7 +13,6 @@ import org.finos.calm.domain.exception.NamespaceNotFoundException;
 import org.finos.calm.store.ArchitectureStore;
 import org.finos.calm.store.PageRequest;
 import org.finos.calm.store.util.MongoVersionDocumentStore;
-import org.finos.calm.store.util.NamespaceGuard;
 
 import java.util.List;
 
@@ -69,19 +68,19 @@ public class MongoArchitectureStore implements ArchitectureStore {
 
     @Override
     public List<NamespaceResourceSummary> getArchitecturesForNamespace(String namespace, PageRequest page) throws NamespaceNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         return documentStore.listSummariesPaged(namespace, page);
     }
 
     @Override
     public boolean architectureExists(String namespace, int architectureId) throws NamespaceNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         return documentStore.headerExists(namespace, architectureId);
     }
 
     @Override
     public Architecture createArchitectureForNamespace(Architecture architecture) throws NamespaceNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, architecture.getNamespace());
+        namespaceStore.requireNamespace(architecture.getNamespace());
 
         // Parsed before the counter is drawn and before anything is written, so malformed
         // JSON can't leave a header behind with no version to go with it.
@@ -158,7 +157,7 @@ public class MongoArchitectureStore implements ArchitectureStore {
     }
 
     private void requireArchitecture(Architecture architecture) throws NamespaceNotFoundException, ArchitectureNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, architecture.getNamespace());
+        namespaceStore.requireNamespace(architecture.getNamespace());
         if (!documentStore.headerExists(architecture.getNamespace(), architecture.getId())) {
             throw new ArchitectureNotFoundException();
         }
