+++ calm-hub/src/main/java/org/finos/calm/store/mongo/MongoFlowStore.java
@@ -14,7 +14,6 @@ import org.finos.calm.domain.namespaces.NamespaceResourceSummary;
 import org.finos.calm.store.FlowStore;
 import org.finos.calm.store.PageRequest;
 import org.finos.calm.store.util.MongoVersionDocumentStore;
-import org.finos.calm.store.util.NamespaceGuard;
 
 import java.util.List;
 
@@ -64,13 +63,13 @@ public class MongoFlowStore implements FlowStore {
 
     @Override
     public List<NamespaceResourceSummary> getFlowsForNamespace(String namespace) throws NamespaceNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         return documentStore.listSummariesPaged(namespace, PageRequest.UNPAGED);
     }
 
     @Override
     public Flow createFlowForNamespace(CreateFlowRequest flowRequest, String namespace) throws NamespaceNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
 
         // Parsed before the counter is drawn and before anything is written, so malformed
         // JSON can't leave a header behind with no version to go with it.
@@ -139,7 +138,7 @@ public class MongoFlowStore implements FlowStore {
     }
 
     private void requireFlow(Flow flow) throws NamespaceNotFoundException, FlowNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, flow.getNamespace());
+        namespaceStore.requireNamespace(flow.getNamespace());
         if (!documentStore.headerExists(flow.getNamespace(), flow.getId())) {
             throw new FlowNotFoundException();
         }
