+++ calm-hub/src/main/java/org/finos/calm/store/nitrite/NitriteInterfaceStore.java
@@ -16,7 +16,6 @@ import org.finos.calm.domain.interfaces.NamespaceInterfaceSummary;
 import org.finos.calm.domain.namespaces.NamespaceResourceSummary;
 import org.finos.calm.store.InterfaceStore;
 import org.finos.calm.store.PageRequest;
-import org.finos.calm.store.util.NamespaceGuard;
 import org.finos.calm.store.util.NitriteVersionDocumentStore;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -63,7 +62,7 @@ public class NitriteInterfaceStore implements InterfaceStore {
 
     @Override
     public List<NamespaceInterfaceSummary> getInterfacesForNamespace(String namespace) throws NamespaceNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         return documentStore.listSummariesPaged(namespace, PageRequest.UNPAGED).stream()
                 .map(NitriteInterfaceStore::toInterfaceSummary)
                 .toList();
@@ -77,7 +76,7 @@ public class NitriteInterfaceStore implements InterfaceStore {
     @Override
     public CalmInterface createInterfaceForNamespace(CreateInterfaceRequest createInterfaceRequest, String namespace) throws NamespaceNotFoundException {
         CalmInterface createdInterface = new CalmInterface(createInterfaceRequest);
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         validateInterfaceJson(createInterfaceRequest.getInterfaceJson());
 
         int id = counterStore.getNextInterfaceSequenceValue();
@@ -110,7 +109,7 @@ public class NitriteInterfaceStore implements InterfaceStore {
 
     @Override
     public CalmInterface createInterfaceForVersion(CreateInterfaceRequest interfaceRequest, String namespace, Integer interfaceId, String version) throws NamespaceNotFoundException, InterfaceNotFoundException, InterfaceVersionExistsException {
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         validateInterfaceJson(interfaceRequest.getInterfaceJson());
         requireInterfaceExists(namespace, interfaceId);
 
@@ -155,7 +154,7 @@ public class NitriteInterfaceStore implements InterfaceStore {
     }
 
     private void requireInterface(String namespace, Integer interfaceId) throws NamespaceNotFoundException, InterfaceNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         requireInterfaceExists(namespace, interfaceId);
     }
 }
