+++ calm-hub/src/main/java/org/finos/calm/store/nitrite/NitritePatternStore.java
@@ -15,7 +15,6 @@ import org.finos.calm.domain.pattern.CreatePatternRequest;
 import org.finos.calm.domain.namespaces.NamespaceResourceSummary;
 import org.finos.calm.store.PageRequest;
 import org.finos.calm.store.PatternStore;
-import org.finos.calm.store.util.NamespaceGuard;
 import org.finos.calm.store.util.NitriteVersionDocumentStore;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -69,13 +68,13 @@ public class NitritePatternStore implements PatternStore {
 
     @Override
     public List<NamespaceResourceSummary> getPatternsForNamespace(String namespace, PageRequest page) throws NamespaceNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         return documentStore.listSummariesPaged(namespace, page);
     }
 
     @Override
     public Pattern createPatternForNamespace(CreatePatternRequest patternRequest, String namespace) throws NamespaceNotFoundException, JsonParseException {
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         validatePatternJson(patternRequest.getPatternJson());
 
         int id = counterStore.getNextPatternSequenceValue();
@@ -116,7 +115,7 @@ public class NitritePatternStore implements PatternStore {
 
     @Override
     public Pattern createPatternForVersion(Pattern pattern) throws NamespaceNotFoundException, PatternNotFoundException, PatternVersionExistsException {
-        NamespaceGuard.requireNamespace(namespaceStore, pattern.getNamespace());
+        namespaceStore.requireNamespace(pattern.getNamespace());
         validatePatternJson(pattern.getPatternJson());
         requirePatternExists(pattern);
 
@@ -134,7 +133,7 @@ public class NitritePatternStore implements PatternStore {
 
     @Override
     public Pattern updatePatternForVersion(Pattern pattern) throws NamespaceNotFoundException, PatternNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, pattern.getNamespace());
+        namespaceStore.requireNamespace(pattern.getNamespace());
         validatePatternJson(pattern.getPatternJson());
         requirePatternExists(pattern);
 
@@ -184,7 +183,7 @@ public class NitritePatternStore implements PatternStore {
     }
 
     private void requirePattern(Pattern pattern) throws NamespaceNotFoundException, PatternNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, pattern.getNamespace());
+        namespaceStore.requireNamespace(pattern.getNamespace());
         requirePatternExists(pattern);
     }
 }
