+++ calm-hub/src/main/java/org/finos/calm/store/nitrite/NitriteStandardStore.java
@@ -15,7 +15,6 @@ import org.finos.calm.domain.standards.CreateStandardRequest;
 import org.finos.calm.domain.namespaces.NamespaceResourceSummary;
 import org.finos.calm.store.PageRequest;
 import org.finos.calm.store.StandardStore;
-import org.finos.calm.store.util.NamespaceGuard;
 import org.finos.calm.store.util.NitriteVersionDocumentStore;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -65,14 +64,14 @@ public class NitriteStandardStore implements StandardStore {
 
     @Override
     public List<NamespaceResourceSummary> getStandardsForNamespace(String namespace) throws NamespaceNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         return documentStore.listSummariesPaged(namespace, PageRequest.UNPAGED);
     }
 
     @Override
     public Standard createStandardForNamespace(CreateStandardRequest createStandardRequest, String namespace) throws NamespaceNotFoundException {
         Standard createdStandard = new Standard(createStandardRequest);
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         validateStandardJson(createStandardRequest.getStandardJson());
 
         int id = counterStore.getNextStandardSequenceValue();
@@ -105,7 +104,7 @@ public class NitriteStandardStore implements StandardStore {
 
     @Override
     public Standard createStandardForVersion(CreateStandardRequest standardRequest, String namespace, Integer standardId, String version) throws NamespaceNotFoundException, StandardNotFoundException, StandardVersionExistsException {
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         validateStandardJson(standardRequest.getStandardJson());
         requireStandardExists(namespace, standardId);
 
@@ -151,7 +150,7 @@ public class NitriteStandardStore implements StandardStore {
     }
 
     private void requireStandard(String namespace, Integer standardId) throws NamespaceNotFoundException, StandardNotFoundException {
-        NamespaceGuard.requireNamespace(namespaceStore, namespace);
+        namespaceStore.requireNamespace(namespace);
         requireStandardExists(namespace, standardId);
     }
 }
