+++ calm-hub/src/test/java/org/finos/calm/store/mongo/TestMongoInterfaceStoreShould.java
@@ -42,6 +42,7 @@ import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.doCallRealMethod;
 
 /**
  * Store-level tests for the header/version shape. Document mechanics are covered by
@@ -76,7 +77,8 @@ public class TestMongoInterfaceStoreShould {
     private static final String VALID_JSON = "{\"test\": \"test\"}";
 
     @BeforeEach
-    void setup() {
+    void setup() throws NamespaceNotFoundException {
+        doCallRealMethod().when(namespaceStore).requireNamespace(anyString());
         headerCollection = Mockito.mock(DocumentMongoCollection.class);
         versionCollection = Mockito.mock(DocumentMongoCollection.class);
